fiks sidebar animasjon
