+++ src/components/ui/input/IconButton.module.css
@@ -20,7 +20,6 @@
   height: var(--size-icon-button-medium);
 }
 
-
 .lg {
   width: var(--size-icon-large);
   height: var(--size-icon-large);
