+++ src/utils/animation.ts
@@ -1,0 +1,2 @@
+// shared exponential easing for UI transitions
+export const EASE_OUT_EXPO = [0.16, 1, 0.3, 1] as const;
